fix(ui): useEnforceScheduleLimit hook created
